+++ tests/modules/test_consoleStatusReporter.py
@@ -1,10 +1,12 @@
+import sys
 import time
 
 from bzt.engine import Provisioning, ScenarioExecutor
 from bzt.modules.jmeter import JMeterExecutor
 from bzt.modules.provisioning import Local
-from bzt.modules.console import ConsoleStatusReporter
+from bzt.modules.console import ConsoleStatusReporter, ConsoleScreen, GUIScreen
 from bzt.modules.aggregator import DataPoint, KPISet
+from bzt.utils import is_windows
 from tests import BZTestCase, r, rc
 from tests.mocks import EngineEmul
 
@@ -89,3 +91,25 @@ class TestConsoleStatusReporter(BZTestCase):
         obj.check()
         obj.shutdown()
         obj.post_process()
+
+    def test_screen(self):
+        isatty = sys.stdout.isatty
+        sys.stdout.isatty = lambda: True
+        obj = ConsoleStatusReporter()
+        obj.settings["screen"] = "console"
+        if not is_windows():
+            self.assertIsInstance(obj._get_screen(), ConsoleScreen)
+        else:
+            self.assertIsInstance(obj._get_screen(), GUIScreen)
+        sys.stdout.isatty = isatty
+
+    def test_screen_invalid(self):
+        isatty = sys.stdout.isatty
+        sys.stdout.isatty = lambda: True
+        obj = ConsoleStatusReporter()
+        obj.settings["screen"] = "invalid"
+        if not is_windows():
+            self.assertIsInstance(obj._get_screen(), ConsoleScreen)
+        else:
+            self.assertIsInstance(obj._get_screen(), GUIScreen)
+        sys.stdout.isatty = isatty